Lecture 349 - Animation States & Transitions & 350 - Implement 2D Collision - Done.
The animation states and transtion was straight forward and now the idle, run and jump animations work and are apart of the animator state machine.

I did a quiz and got 3 out of 5 correct.

The 2D collision stuff was quite intersting. The tile map collider, pyshics shape tool and composite collider were all new to me.