Lecture 354 - Instructor Hangout 6.1 - Done.
This helped explain some of the code. I get it a bit more now.
Also Ben talked about the concept of Calculate versus Cache or Performance versus Easy to read code.

Calculate is fast because processors these days are really fast. Cache is slow because memory isn't that quick at all. So if you want things to be quick then don't rely too much on cache.