Lecture 371 - Ladder Tweaks - Done.
The ladder bug was fixed by Rick his own slightly different way but then he fixed the bug where the player falls down the ladder slowly but in order to have the same shit as him I had to change my code. This was quite the shame.

The player still doesn't move down the ladder but it now is able to at any point on the ladder and stand their likes its the floor.