Lecture 380 - Level Flow Layout - Improved.
I improved the levels made in this lecture and made it a bit fairer to play through.

I also found that the enemy box collider is hitting the player if they are on the lower square next to them and that counts as an attack.
When I tried changing the on trigger enter collider to a capsule collider it made it so that no attack worked and the enemy couldn't kill the player. I'll have to return to this later on when the lectures are over with. Maybe they spot the bug later, if not I'll have to fix it.