Lecture 382 - Start & Success Screens - Done.
I made the start and win screens mostly. My idea for the win screen has remained incomplete due to my inability to set up the set active UI enable crap.
Also the state driven camera is a mess too.

I'll have to come back one day and ethier fix this or scrap it and do something else with the menu. Maybe a similar thing but without the camera movement. Just a menu that fades in over a set amount of time.