Lecture 330(Should be 388) - Remembering Pickups - Done/Broken
The aim was to make it so that the coin pickups don't re-spawn when you lose a life but that doesn't seem to work as well several other things in this game. The bugs just keep piling up and they aren't focused on this at the moment. Bringing in new students/customers is their priority and making new courses for the current customer/student fan base can spend on.